moved prepare content to publish
